refactor: information에 Exception 수정, forum에 Exception 수정
- ErrorException을 InfoArticleNotFoundException으로 이름 변경
- Forum에 ForumNotFoundException 추가
