+++ src/main/java/kusuri12/teens_be/domain/comment/exception/ForumNotFoundException.java
@@ -1,7 +1,11 @@
 package kusuri12.teens_be.domain.comment.exception;
 
-public class ForumNotFoundException extends RuntimeException {
-  public ForumNotFoundException(String message) {
-    super(message);
-  }
+import kusuri12.teens_be.global.error.exception.ErrorCode;
+import kusuri12.teens_be.global.error.exception.TeensException;
+
+public class ForumNotFoundException extends TeensException {
+
+    public static final TeensException EXCEPTION = new ForumNotFoundException();
+
+    public ForumNotFoundException() { super(ErrorCode.FORUM_NOT_FOUND); }
 }
