+++ src/main/java/kusuri12/teens_be/domain/information/exception/InfoArticleNotFoundException.java
@@ -3,8 +3,10 @@ package kusuri12.teens_be.domain.information.exception;
 import kusuri12.teens_be.global.error.exception.ErrorCode;
 import kusuri12.teens_be.global.error.exception.TeensException;
 
-public class informayion extends TeensException {
-  public informayion(ErrorCode errorCode) {
-    super(errorCode);
+public class InfoArticleNotFoundException extends TeensException {
+    public static final TeensException EXCEPTION = new InfoArticleNotFoundException();
+
+    public InfoArticleNotFoundException() {
+      super(ErrorCode.INFO_ARTICLE_NOT_FOUND);
   }
 }
