+++ src/main/java/kusuri12/teens_be/domain/information/service/InfoArticleService.java
@@ -1,11 +1,12 @@
 package kusuri12.teens_be.domain.information.service;
 
-import kusuri12.teens_be.domain.information.dto.InfoArticleDto;
+import kusuri12.teens_be.domain.information.presentation.dto.InfoArticleDto;
 import kusuri12.teens_be.domain.information.domain.InfoArticle;
-import kusuri12.teens_be.domain.information.repository.InfoArticleRepository;
+import kusuri12.teens_be.domain.information.domain.repository.InfoArticleRepository;
 import kusuri12.teens_be.domain.user.domain.User;
 import kusuri12.teens_be.domain.user.domain.repository.UserRepository;
-import kusuri12.teens_be.domain.information.exception.ErrorException;
+import kusuri12.teens_be.domain.information.exception.InfoArticleNotFoundException;
+import kusuri12.teens_be.domain.user.exception.UserNotFoundException;
 import kusuri12.teens_be.global.error.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -38,7 +39,7 @@ public class InfoArticleService {
 
     public InfoArticleDto.InfoArticleDetailResponse getInfoArticleDetail(Long articleId) {
         InfoArticle article = infoArticleRepository.findById(articleId)
-                .orElseThrow(() -> new ErrorException(ErrorCode.INFO_ARTICLE_NOT_FOUND));
+                .orElseThrow(() -> InfoArticleNotFoundException.EXCEPTION);
 
         return InfoArticleDto.InfoArticleDetailResponse.builder()
                 .id(article.getId())
@@ -53,7 +54,7 @@ public class InfoArticleService {
     @Transactional
     public void createInfoArticle(InfoArticleDto.CreateInfoArticleRequest request) {
         User user = userRepository.findById(request.getUserId())
-                .orElseThrow(() -> new ErrorException(ErrorCode.USER_NOT_FOUND));
+                .orElseThrow(() -> UserNotFoundException.EXCEPTION);
 
         InfoArticle article = InfoArticle.builder()
                 .title(request.getTitle())
