+++ src/main/java/kusuri12/teens_be/global/error/exception/ErrorCode.java
@@ -38,7 +38,10 @@ public enum ErrorCode {
     INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "Internal Server Error"),
 
     // 정보 게시판 관련
-    INFO_ARTICLE_NOT_FOUND(HttpStatus.NOT_FOUND, "해당 게시글을 찾을 수 없습니다.");;
+    INFO_ARTICLE_NOT_FOUND(HttpStatus.NOT_FOUND, "해당 게시글을 찾을 수 없습니다."),
+
+    // 티키타카 게시판 관련
+    FORUM_NOT_FOUND(HttpStatus.NOT_FOUND, "해당 게시글을 찾을 수 없습니다.");
 
     private final HttpStatus status;
     private final String message;
